Reorganize source around XML and JSON folders
